add translation package including section registration
