+++ package.json
@@ -94,6 +94,7 @@
 		"./temporary-file": "./dist-cms/packages/core/temporary-file/index.js",
 		"./themes": "./dist-cms/packages/core/themes/index.js",
 		"./tiny-mce": "./dist-cms/packages/tiny-mce/index.js",
+		"./translation": "./dist-cms/packages/translation/index.js",
 		"./tree": "./dist-cms/packages/core/tree/index.js",
 		"./ufm": "./dist-cms/packages/ufm/index.js",
 		"./user-change-password": "./dist-cms/packages/user/change-password/index.js",
