--- src/packages/dictionary/section/constants.ts
@@ -1,1 +1,0 @@
-export const UMB_DICTIONARY_SECTION_ALIAS = 'Umb.Section.Translation';
