--- src/packages/dictionary/section/index.ts
@@ -1,1 +1,0 @@
-export { UMB_DICTIONARY_SECTION_ALIAS } from './constants.js';
