+++ src/packages/translation/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_TRANSLATION_SECTION_ALIAS = 'Umb.Section.Translation';
