+++ tsconfig.json
@@ -122,6 +122,7 @@ DON'T EDIT THIS FILE DIRECTLY. It is generated by /devops/tsconfig/index.js
 			"@umbraco-cms/backoffice/temporary-file": ["./src/packages/core/temporary-file/index.ts"],
 			"@umbraco-cms/backoffice/themes": ["./src/packages/core/themes/index.ts"],
 			"@umbraco-cms/backoffice/tiny-mce": ["./src/packages/tiny-mce/index.ts"],
+			"@umbraco-cms/backoffice/translation": ["./src/packages/translation/index.ts"],
 			"@umbraco-cms/backoffice/tree": ["./src/packages/core/tree/index.ts"],
 			"@umbraco-cms/backoffice/ufm": ["./src/packages/ufm/index.ts"],
 			"@umbraco-cms/backoffice/user-change-password": ["./src/packages/user/change-password/index.ts"],
